server/requirements.txt
Flask
Flask-Cors
openai
python-dotenv
markdown2
requests